+++ test/grammar/condition_floating_point_test.rb
@@ -1,34 +1,9 @@
 require File.expand_path('base_test.rb', File.dirname(__FILE__))
 
 module Grammar
-  class ConditionTest < BaseTest
-    def test_select_where_parseable
-      assert_ast_sql_equal "select * from table1 where col1 = col1",
-        Ast::SelectStatement[
-          :subquery => Ast::Subquery[
-            :query_block => Ast::QueryBlock[
-              :select_list => Ast::Array[
-                Ast::Identifier[:name => '*']
-              ],
-              :select_sources => Ast::Array[
-                Ast::TableReference[
-                 :table_name => Ast::Identifier[:name => 'table1']
-                ]
-              ],
-              :where_clause => Ast::WhereClause[
-                :condition => Ast::SimpleComparisonCondition[
-                  :left => Ast::Identifier[:name => 'col1'],
-                  :op => '=',
-                  :right => Ast::Identifier[:name => 'col1']
-                ]
-              ]
-            ]
-          ]
-        ]
-    end
-
-    def test_select_where_with_literal_textparseable
-      assert_ast_sql_equal "select * from table1 where col1 = 'abc'",
+  class ConditionFloatingPointTest < BaseTest
+    def test_select_where_floating_point_condition_nan_parseable
+      assert_ast_sql_equal "select * from table1 where col1 is nan",
         Ast::SelectStatement[
           :subquery => Ast::Subquery[
             :query_block => Ast::QueryBlock[
@@ -41,10 +16,10 @@ module Grammar
                 ]
               ],
               :where_clause => Ast::WhereClause[
-                :condition => Ast::SimpleComparisonCondition[
-                  :left => Ast::Identifier[:name => 'col1'],
-                  :op => '=',
-                  :right => Ast::TextLiteral[:value => 'abc']
+                :condition => Ast::FloatingPointCondition[
+                  :target => Ast::Identifier[:name => 'col1'],
+                  :is => Ast::Keyword[:name => 'is'],
+                  :value => Ast::Keyword[:name => 'nan'],
                 ]
               ]
             ]
@@ -52,8 +27,8 @@ module Grammar
         ]
     end
 
-    def test_select_where_with_literal_number_parseable
-      assert_ast_sql_equal "select * from table1 where col1 = -1",
+    def test_select_where_floating_point_condition_with_not_parseable
+      assert_ast_sql_equal "select * from table1 where col1 is not nan",
         Ast::SelectStatement[
           :subquery => Ast::Subquery[
             :query_block => Ast::QueryBlock[
@@ -66,15 +41,44 @@ module Grammar
                 ]
               ],
               :where_clause => Ast::WhereClause[
-                :condition => Ast::SimpleComparisonCondition[
-                  :left => Ast::Identifier[:name => 'col1'],
-                  :op => '=',
-                  :right => Ast::NumberLiteral[:value => '-1']
+                :condition => Ast::FloatingPointCondition[
+                  :target => Ast::Identifier[:name => 'col1'],
+                  :is => Ast::Keyword[:name => 'is'],
+                  :not => Ast::Keyword[:name => 'not'],
+                  :value => Ast::Keyword[:name => 'nan'],
                 ]
               ]
             ]
           ]
         ]
     end
+
+    def test_select_where_floating_point_condition_infinit_parseable
+      assert_ast_sql_equal "select * from table1 where col1 is not infinite",
+        Ast::SelectStatement[
+          :subquery => Ast::Subquery[
+            :query_block => Ast::QueryBlock[
+              :select_list => Ast::Array[
+                Ast::Identifier[:name => '*']
+              ],
+              :select_sources => Ast::Array[
+                Ast::TableReference[
+                  :table_name => Ast::Identifier[:name => 'table1']
+                ]
+              ],
+              :where_clause => Ast::WhereClause[
+                :condition => Ast::FloatingPointCondition[
+                  :target => Ast::Identifier[:name => 'col1'],
+                  :is => Ast::Keyword[:name => 'is'],
+                  :not => Ast::Keyword[:name => 'not'],
+                  :value => Ast::Keyword[:name => 'infinite'],
+                ]
+              ]
+            ]
+          ]
+        ]
+    end
+
+
   end
 end